First cut at storing line number when tokenizing
It might be easier to read the file line by line but I figure that
will complicate it later if I want to have tokens that span multiple
lines, such as string literals. This change also moves
whitespace/newline detection earlier in the loop to make it clearer.
Also reformatted entire file.